backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
firebase-admin==6.2.0
APScheduler==3.10.4
python-dotenv==1.0.0